+++ packages/components-angular/src/app/examples/inline-notification-example-action-button.component.ts
@@ -1,21 +1,21 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
 
 @Component({
-  selector: 'page-banner-inline-example-action-button',
+  selector: 'page-inline-notification-example-action-button',
   template: `
-    <p-banner-inline
-      [heading]="'Some banner-inline heading'"
-      [description]="'Some banner-inline description.'"
+    <p-inline-notification
+      [heading]="'Some inline-notification heading'"
+      [description]="'Some inline-notification description.'"
       [actionLabel]="'Retry'"
       [actionIcon]="'reset'"
       [actionLoading]="isLoading"
       (action)="onAction()"
-    ></p-banner-inline>
+    ></p-inline-notification>
     <p-button (click)="onAction()">Reset</p-button>
   `,
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class BannerInlineExampleActionButtonComponent {
+export class InlineNotificationExampleActionButtonComponent {
   isLoading = false;
 
   onAction() {