+++ packages/components/CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### Added
 - `Marque` uses `webp` images for browsers that support it
-- `Banner Inline`
+- `Inline Notification`
 - `Icon` now supports `success` for `name` property
 
 #### Fixed